[CXF-6988] Add more systests for CDI, including OpenWebBeans.
