+++ systests/cdi/base/src/main/resources/META-INF/beans.xml
