+++ systests/cdi/cdi-no-app-weld/src/test/resources/META-INF/services/org.jboss.weld.environment.Container
