+++ systests/cdi/cdi-no-app-weld/src/test/resources/jaxrs_cdi/WEB-INF/web.xml
