+++ systests/cdi/cdi-weld/src/test/resources/META-INF/services/org.jboss.weld.environment.Container
@@ -1,0 +1,1 @@
+org.apache.cxf.systest.jaxrs.cdi.WeldDiscoverableContainer
