--- systests/cdi/src/test/java/org/apache/cxf/systest/jaxrs/cdi/Book.java
@@ -1,54 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.cxf.systest.jaxrs.cdi;
-
-import javax.validation.constraints.NotNull;
-
-public class Book {
-    @NotNull private String name;
-    private String id;
-    
-    public Book() {
-    }
-    
-    public Book(String id) {
-        this.id = id;
-    }
-    
-    public Book(String name, String id) {
-        this.name = name;
-        this.id = id;
-    }
-    
-    public void setName(String n) {
-        name = n;
-    }
-
-    public String getName() {
-        return name;
-    }
-    
-    public void setId(String i) {
-        id = i;
-    }
- 
-    public String getId() {
-        return id;
-    }
-}
