--- systests/cdi/src/test/java/org/apache/cxf/systest/jaxrs/cdi/jetty/JettyEmbeddedTest.java
@@ -1,49 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.cxf.systest.jaxrs.cdi.jetty;
-
-import org.apache.cxf.jaxrs.model.AbstractResourceInfo;
-import org.apache.cxf.systest.jaxrs.cdi.AbstractCDITest;
-import org.junit.BeforeClass;
-import org.junit.Ignore;
-
-public class JettyEmbeddedTest extends AbstractCDITest {  
-    @Ignore
-    public static class EmbeddedJettyServer extends AbstractJettyServer {
-        public static final int PORT = allocatePortAsInt(EmbeddedJettyServer.class);
-
-        public EmbeddedJettyServer() {
-            super("/", PORT);
-        }
-    }
-    
-    @BeforeClass
-    public static void startServers() throws Exception {
-        AbstractResourceInfo.clearAllMaps();
-        //keep out of process due to stack traces testing failures
-        assertTrue("server did not launch correctly", launchServer(EmbeddedJettyServer.class, true));
-        createStaticBus();
-    }
-    
-    @Override
-    protected int getPort() {
-        return EmbeddedJettyServer.PORT;
-    }
-}
